requirements.txt
fastapi==0.109.0
passlib==1.7.4
pydantic==1.10.12
pymongo==4.6.1
python_jose==3.3.0
yfinance==0.2.36
